Use mixins for filter value
